Before: #69 


After:

```
bench_rctree            time:   [1.3840 us 1.3905 us 1.3982 us]                          
                        change: [-24.090% -22.771% -21.430%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
  6 (6.00%) high severe

bench_rctree_get_joints time:   [205.40 ns 206.45 ns 207.83 ns]                                    
                        change: [-30.009% -29.069% -28.105%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe

bench_rctree_set_joints time:   [191.14 ns 191.40 ns 191.68 ns]                                    
                        change: [-23.614% -22.624% -21.666%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) high mild
  9 (9.00%) high severe

bench_rctree_concurrent_set_joints_1000                                                                             
                        time:   [1.5055 ms 1.5074 ms 1.5093 ms]
                        change: [-46.751% -46.430% -46.145%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe

bench_rctree_concurrent_get_joints_1000                                                                             
                        time:   [1.0934 ms 1.0947 ms 1.0961 ms]
                        change: [-60.615% -60.376% -60.150%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe

bench_rctree_concurrent_set_get_joints_1000                                                                             
                        time:   [1.3913 ms 1.3926 ms 1.3938 ms]
                        change: [-50.261% -50.001% -49.785%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) low severe
  2 (2.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

bench_rctree_ik         time:   [6.5265 us 6.5845 us 6.6759 us]                             
                        change: [-12.322% -11.582% -10.704%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  4 (4.00%) low severe
  2 (2.00%) low mild
  4 (4.00%) high mild
  9 (9.00%) high severe
```